Un-Rose 2
Separate main module into 2 modules. One containing only CLI stuffs and
the other containing the DateTimeOperator class.

Migrate tests from `rose test-battery` to Pythonic tests.

Add `isodatetime` command.
